requirements.txt
amqp==5.0.6
asgiref==3.3.1
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.7
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-chartjs==2.2.1
django-extensions==3.1.2
django-timezone-field==4.1.2
idna==2.10
kombu==5.0.2
prompt-toolkit==3.0.18
pycparser==2.20
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
rabbitmq==0.2.0
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
